mind_genrec/requirements.txt
faiss-cpu>=1.7,<2.0
fastapi>=0.110,<1.0
numpy>=1.26,<3.0
pydantic>=2.6,<3.0
PyYAML>=6.0,<7.0
redis>=5.0,<6.0
torch>=2.2,<3.0
uvicorn>=0.29,<1.0
